Allow for specifying polyfillOptions in order to disable the triggering of DOMNode insertion events
If the developer does not want to rely on asynchronous events being fired when the DOM has been manipulated, he has to be able to disable the event emitting.
In order to enable the developer the polyfill has been extended to make use of an options hash, which currently includes the boolean options "DOMNodeInsertionEvents".
By default the polyfill will keep its original behavior, thereby mitigating asynchronous events.
When DOMNodeInsertionEvents is set to false, the MutationObserver callback won't trigger any attach callbacks for inserted DOM nodes. Further, the "DOMNodeInserted" event listeners won't be attached as a fallback.