Expose helper function to allow for synchronous activation of Custom Elements
When the asynchronous insertion events are suppressed the developer needs to have an option to "activate" his components and trigger the create- and attach-callbacks.
Therefore we are exposing the internal helper method for this specific use-case. The method activateComponentsInNode would require a rootNode to be provided to make a deep search and verify on the node and its childNodes.